# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Recruit API",
        "description": "採用広報サイト + 応募管理ダッシュボード API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/news": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "お知らせ一覧（公開）",
                "operationId": "b83ff020eb45141b07c365e50a4b27bd",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "お知らせ一覧"
                    }
                }
            }
        },
        "/news/{slug}": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "お知らせ詳細（公開）",
                "operationId": "d6b1b57588ce607cc07682593acedd4d",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "お知らせ詳細"
                    },
                    "404": {
                        "description": "見つからない"
                    }
                }
            }
        },
        "/admin/news": {
            "get": {
                "tags": [
                    "Admin/News"
                ],
                "summary": "お知らせ一覧（管理）",
                "operationId": "073330a93887434c288427e09f53a8c4",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "お知らせ一覧"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin/News"
                ],
                "summary": "お知らせ作成",
                "operationId": "639c0a5b847dee03515137a040a2fc16",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "作成成功"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/news/{news}": {
            "put": {
                "tags": [
                    "Admin/News"
                ],
                "summary": "お知らせ更新",
                "operationId": "beeedfce62bfc6e32c060c110ff9af15",
                "parameters": [
                    {
                        "name": "news",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "更新成功"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin/News"
                ],
                "summary": "お知らせ削除",
                "operationId": "031750e603ff885ed61d79e5ca86c530",
                "parameters": [
                    {
                        "name": "news",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "削除成功"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Sanctum Token",
                "scheme": "bearer"
            }
        }
    },
...
}